Refactor test and colocate test in one file
